Allow smart builds to reuse the shader cache
Previously if a mod had content which might contain shaders (.upk or
.umap files in the Content folder) we would always compile shaders.
However, many builds are just code rather than content changes, hence
this step is often redundant.

Now we check the last write time on the shader cache from the deployed
(installed, in the game) build of the mod, and compare it to the most
recent last write time of content which might contain shaders. If the
shader cache is more recent, we take a copy of the previous shader cache
rather than rebuilding it. This is much faster.