Search for the highlander in game mods and Steam
Previously we presumed the community highlander, if required for a
--highlander build, would be found in the relevant game edition's
Mods folder. Now we search both that and Steam workshop folders,
looking for a copy of the relevant highlander.

If we find one, great. If we find none, we report an error. If we find
several, we report an error and the locations at which we found
highlanders. It's up to the user to remove the ones they don't want,
or rename their .XComMod files to hide them from both the game and this
tool.